pyproject.toml
[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.poetry]
name = "vizme-backend"
version = "0.1.0"
description = ""
authors = ["pavivin <pavivin@yandex.ru>"]
readme = "README.md"
packages = [{include = "vizme_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = {extras = ["asyncpg"], version = "^2.0.9"}
httpx = "^0.24.0"
celery = {extras = ["redis"], version = "^5.2.7"}
orjson = "^3.8.10"
pydantic = "^1.10.7"
alembic = "^1.10.3"
pytest-asyncio = "^0.21.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
factory-boy = "^3.2.1"
async-factory-boy = "^1.0.0"
psycopg2-binary = "^2.9.6"
python-socketio = "^5.8.0"
aiohttp = "^3.8.4"
fastapi = "^0.95.1"
sqladmin = "^0.10.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
